Removing useSessionValidation hook and placing the validation logic from token into sessionPRovider component
