+++ src/app/[locale]/_component/sessionProvider.tsx
@@ -1,10 +1,13 @@
 'use client';
 
-import { useRouter, usePathname } from 'next/navigation';
+import { useRouter, usePathname, redirect } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { LOGIN_ROUTES, PUBBLIC_ROUTES, ROUTES } from '../_utils/routes';
 import useToken from '../_hooks/useToken';
-import useSessionValidation from '../_hooks/useSessionValidation';
+import { isBrowser } from '../_utils/common';
+import { SpidValueInJWT } from '../_model/JWTUser';
+import { extractToken, userFromJwtToken } from '../_utils/jwt';
+import { storageTokenOps, storageUserOps } from '../_utils/storage';
 import Loader from './loader/loader';
 
 type LoginStatusIdle = {
@@ -24,11 +27,43 @@ export type LoginStatus = LoginStatusIdle | LoginStatusAuthorized | LoginStatusN
 const SessionProviderComponent = ({ children }: { readonly children: React.ReactNode }) => {
   const [loginStatus, setLoginStatus] = useState<LoginStatus>({ status: 'IDLE' });
   const { isTokenValid, removeToken } = useToken();
-  useSessionValidation();
   const router = useRouter();
   const pathName = usePathname();
 
   const cleanPath = (path: string): string => path.replace(/^(\/(en|it))\/(.*)$/, '');
+
+  const token = isBrowser() ? extractToken() : undefined;
+  const userFromToken = token ? userFromJwtToken(token) : undefined;
+
+  const L1_JWT_LEVEL: SpidValueInJWT = {
+    value: process.env.NEXT_PUBLIC_JWT_SPID_LEVEL_VALUE_L1,
+  };
+
+  const L2_JWT_LEVEL: SpidValueInJWT = {
+    value: process.env.NEXT_PUBLIC_JWT_SPID_LEVEL_VALUE_L2,
+  };
+
+  const L3_JWT_LEVEL: SpidValueInJWT = {
+    value: process.env.NEXT_PUBLIC_JWT_SPID_LEVEL_VALUE_L3,
+  };
+
+  useEffect(() => {
+    if (token && userFromToken) {
+      storageTokenOps.write(token);
+      storageUserOps.write(userFromToken);
+      switch (userFromToken?.spidLevel) {
+        case L1_JWT_LEVEL.value:
+          redirect(ROUTES.SESSION);
+          break;
+        case L2_JWT_LEVEL.value:
+          redirect(ROUTES.PROFILE);
+          break;
+        case L3_JWT_LEVEL.value:
+          redirect(ROUTES.PROFILE_RESTORE);
+          break;
+      }
+    }
+  }, []);
 
   useEffect(() => {
     if (LOGIN_ROUTES.includes(cleanPath(pathName))) {
@@ -44,15 +79,10 @@ const SessionProviderComponent = ({ children }: { readonly children: React.React
       setLoginStatus({ status: 'NOT_AUTHORIZED' });
       router.push(ROUTES.LOGIN);
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [pathName]);
 
   if (loginStatus.status === 'IDLE' || loginStatus.status === 'NOT_AUTHORIZED') {
-    return (
-      <>
-        <Loader />
-      </>
-    );
+    return <Loader />;
   }
   return <>{children}</>;
 };
